Split Settings page to multiple, other CSS fixes
