+++ client/src/App.css
@@ -33,6 +33,7 @@ ion-select.shorter-select.shorter-select2 {
     min-height: 42px;
 }
 
+
 .floating-error-footer {
     border: none;
     outline: none;
